truck location/info ducks refactor
