+++ src/types/location.js
@@ -1,0 +1,7 @@
+const types = {
+  TRUCK_LOCATION_FETCH_START: "TRUCK_LOCATION_FETCH_START",
+  TRUCK_LOCATION_FETCH_FINISHED: "TRUCK_LOCATION_FETCH_FINISHED",
+  TRUCK_LOCATION_FETCH_ERROR: "TRUCK_LOCATION_FETCH_ERROR",
+};
+
+export default types;
