+++ src/types/truck.js
@@ -2,9 +2,6 @@ const types = {
   TRUCK_INFO_FETCH_START: "TRUCK_INFO_FETCH_START",
   TRUCK_INFO_FETCH_FINISHED: "TRUCK_INFO_FETCH_FINISHED",
   TRUCK_INFO_FETCH_ERROR: "TRUCK_INFO_FETCH_ERROR",
-  TRUCK_LOCATION_FETCH_START: "TRUCK_LOCATION_FETCH_START",
-  TRUCK_LOCATION_FETCH_FINISHED: "TRUCK_LOCATION_FETCH_FINISHED",
-  TRUCK_LOCATION_FETCH_ERROR: "TRUCK_LOCATION_FETCH_ERROR",
 };
 
 export default types;
